Attempt to update live activity through background tasks framework (expanded description included)
However, the background tasks framework is too unreliable for this use-case.
Currently, there are no exposed APIs that allow for live activities to be updated in a timely manner
other than remote push notifications (which doesn't exactly fit this use-case for a timer in airplane mode).

I'll have to wait for apple to update live activities to use timeline entries like their other widget APIs.